refactoring with implementation of new db schema, all registers with words the same for all PMs

+++ mapi/include/RefreshMapiGroup.h
@@ -14,10 +14,10 @@ public:
 
 private:
     Fred* fred;
-    uint32_t hexValue;
+    int64_t hexValue;
     bool firstTime;
     string sequence;
     std::vector<std::string> services;
-    std::vector<uint32_t> oldValues;
+    std::vector<int64_t> oldValues;
     std::vector<std::pair<std::string, std::string>> refreshServices;
 };
